Snake backend controller
This is the backend of the snake game. It recieves the number of players, the path to a world map and where to log the events.
Check worlds/ for examples of world maps
At the end it logs the game status and the game winner.
Then for each turn, it will log n+1 lines, one for each player and a extra line
each player line will have:
1.- A number 1 if turn right, or -1 if turn left, or 0 if moved forward.
2.- A boolean, 1 if the snake ate
 0 otherwise
3.- A boolean, 1 if the snake is playing, 0 if it died :(

The extra line will have two numbers, the row,column coordinate of a food that was spawned
if no food was spawned, then it will have two numbers -1 -1